Split relaunch deadline tests by scenario r?lissyx
