+++ testing/enterprise/manifest-browser.toml
@@ -72,7 +72,17 @@ run-if = [
 
 ["test_felt_harness_prefs.py"]
 
-["test_felt_browser_relaunch_deadline.py"]
+["test_felt_browser_relaunch_deadline_countdown_focus.py"]
+
+["test_felt_browser_relaunch_deadline_infobar_slot.py"]
+
+["test_felt_browser_relaunch_deadline_missing_policies.py"]
+
+["test_felt_browser_relaunch_deadline_removed_bar.py"]
+
+["test_felt_browser_relaunch_deadline_restart.py"]
+
+["test_felt_browser_relaunch_deadline_withdrawn.py"]
 
 ["test_felt_restart_persists_session.py"]
